refactor: 肥大化した定番品・カテゴリ管理コンポーネントを分割
staple-items-sheet.tsx (434行) から SortableStapleCard と
StapleQuickAddSheet を別ファイルへ抽出（→337行）。
category-manager.tsx (259行) から SortableCategoryRow / CategoryRowOverlay を
抽出し、編集行と追加フォームで重複していた色選択 UI を ColorPicker に共通化（→141行）。
挙動は変更なし。
