+++ src/components/category/color-picker.tsx
@@ -1,0 +1,26 @@
+"use client";
+
+export const PRESET_COLORS = [
+  "#ef4444", "#f97316", "#eab308", "#22c55e",
+  "#3b82f6", "#6366f1", "#a855f7", "#ec4899",
+];
+
+interface ColorPickerProps {
+  value: string;
+  onChange: (color: string) => void;
+}
+
+export function ColorPicker({ value, onChange }: ColorPickerProps) {
+  return (
+    <div className="flex gap-1.5">
+      {PRESET_COLORS.map((c) => (
+        <button
+          key={c}
+          onClick={() => onChange(c)}
+          className={`h-6 w-6 rounded-full border-2 ${value === c ? "border-foreground" : "border-transparent"}`}
+          style={{ backgroundColor: c }}
+        />
+      ))}
+    </div>
+  );
+}
